Added map scalibility
Both Visible map size is scalable, so it can now support zooming in and out (Although at the moment this setting is hardcoded). However, buttons will shift position so that movement keys are always cardinal-adjacent to the middle. (Need to add visuals to represent center or emphasize movement keys somehow, or just run map on a low size.)